requirements.txt
discord.py==2.3.2
moviepy==1.0.3
pytube==15.0.0
